requirements.txt
# Django Core
django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# Database
mysqlclient>=2.2.0

# WebSocket
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Celery
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0

# File Storage
minio>=7.2.0
boto3>=1.34.0

# Utils
python-dotenv>=1.0.0
Pillow>=10.0.0
pytz>=2024.1

# Development
pytest>=7.4.0
pytest-django>=4.7.0
factory-boy>=3.3.0
drf-spectacular >= 0.29.0